requirements.txt
numpy==1.23.3
websockets==11.0.3
openai==0.27.8
tiktoken==0.4.0
requests==2.26
uvicorn==0.22.0
matplotlib==3.7.1
scipy==1.10.1
mariadb==1.1.7
pandas==2.0.2
Pillow==9.4.0
gradio==3.23.0
pymongo==4.4.1
python-dotenv==0.10.1
sympy==1.12
retry==0.9.2
pydub==0.25.1
PyPDF2==3.0.1
langchain==0.0.149
fastapi==0.94.0
pylatex==1.4.1